Rename custom local libraries directory, adn added files to ignore
