+++ mtgai/common.py
@@ -1,8 +1,8 @@
-# mtg_common_lib.py
+# common.py
 # Stores common functions for logging, and text file creation, etc.
 
-from lib import mtg_logging_lib as log
-from lib import mtg_data_lib as mtg
+from mtgai import custom_logging as log
+from mtgai import mtg_data as mtg
 
 def Menu():
     log.PrintAndLog("Displaying menu to user")
